# api/handlers/v1/category.go
package v1

import (
	"context"
	"log"
	"net/http"
	"strconv"
	"univer/api/models"
	"univer/internal/entity"

	"github.com/gin-gonic/gin"
	"google.golang.org/protobuf/encoding/protojson"
)

// @Security 		BearerAuth
// @Summary 		Create Category
// @Description 	This API for create a new category for product
// @Tags 			category
// @Produce 		json
// @Accept 			json
// @Param 			order body models.CategoryReq true "Create Category Model"
// @Success			201 {object} models.Category
// @Failure 		400 {object} models.Error
// @Failure 		401 {object} models.Error
// @Failure 		403 {object} models.Error
// @Faulure 		500 {object} models.Error
// @Router 			/v1/category [POST]
func (h *HandlerV1) CreateCategory(c *gin.Context) {
	var (
		body        models.CategoryReq
		jspbMarshal protojson.MarshalOptions
	)
	jspbMarshal.UseProtoNames = true

	ctx, cancel := context.WithTimeout(context.Background(), h.Config.Context.Timeout)
	defer cancel()

	err := c.ShouldBindJSON(&body)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}

	category, err := h.Service.Category().CreateCategory(ctx, &entity.Category{
		Name: body.Name,
	})
	if err != nil {
		c.JSON(http.StatusInternalServerError, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}

	c.JSON(http.StatusCreated, models.Category{
		ID:   category.Id,
		Name: category.Name,
	})
}

// @Security 		BearerAuth
// @Summary 		Update Category
// @Description 	This API for update a category
// @Tags 			category
// @Produce 		json
// @Accept 			json
// @Param 			order body models.Category true "Create Category Model"
// @Success			200 {object} models.Category
// @Failure 		400 {object} models.Error
// @Failure 		401 {object} models.Error
// @Failure 		403 {object} models.Error
// @Faulure 		500 {object} models.Error
// @Router 			/v1/category [PUT]
func (h *HandlerV1) UpdateCategory(c *gin.Context) {
	var (
		body        models.Category
		jspbMarshal protojson.MarshalOptions
	)
	jspbMarshal.UseProtoNames = true

	ctx, cancel := context.WithTimeout(context.Background(), h.Config.Context.Timeout)
	defer cancel()

	err := c.ShouldBindJSON(&body)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}

	category, err := h.Service.Category().UpdateCategory(ctx, &entity.UpdateCategory{
		Id:   body.ID,
		Name: body.Name,
	})
	if err != nil {
		c.JSON(http.StatusInternalServerError, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}

	c.JSON(http.StatusOK, models.Category{
		ID:   category.Id,
		Name: category.Name,
	})
}

// @Security 		BearerAuth
// @Summary 		Delete Category
// @Description 	This API for delete a category with id
// @Tags 			category
// @Produce 		json
// @Accept 			json
// @Param 			id path string true "Category ID"
// @Success			200 {object} string
// @Failure 		401 {object} models.Error
// @Failure 		403 {object} models.Error
// @Failure 		404 {object} models.Error
// @Faulure 		500 {object} models.Error
// @Router 			/v1/category/{id} [DELETE]
func (h *HandlerV1) DeleteCategory(c *gin.Context) {
	var (
		jspbMarshal protojson.MarshalOptions
	)
	jspbMarshal.UseProtoNames = true

	ctx, cancel := context.WithTimeout(context.Background(), h.Config.Context.Timeout)
	defer cancel()

	categoryID := c.Param("id")

	err := h.Service.Category().DeleteCategory(ctx, &entity.DeleteReq{
		Id: categoryID,
	})
	if err != nil {
		c.JSON(http.StatusNotFound, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}

	c.JSON(http.StatusOK, "Category has been deleted succesfully")
}

// @Security 		BearerAuth
// @Summary 		Get Category
// @Description 	This API for getting a category with id
// @Tags 			category
// @Produce 		json
// @Accept 			json
// @Param 			id path string true "Category ID"
// @Success			200 {object} models.Category
// @Failure 		404 {object} models.Error
// @Failure 		401 {object} models.Error
// @Failure 		403 {object} models.Error
// @Faulure 		500 {object} models.Error
// @Router 			/v1/category/{id} [GET]
func (h *HandlerV1) GetCategory(c *gin.Context) {
	var (
		jspbMarshal protojson.MarshalOptions
	)
	jspbMarshal.UseProtoNames = true

	ctx, cancel := context.WithTimeout(context.Background(), h.Config.Context.Timeout)
	defer cancel()

	categoryID := c.Param("id")

	filter := map[string]string{
		"id": categoryID,
	}

	category, err := h.Service.Category().GetCategory(ctx, &entity.GetReq{
		Filter: filter,
	})
	if err != nil {
		c.JSON(http.StatusNotFound, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}

	c.JSON(http.StatusOK, models.Category{
		ID:   categoryID,
		Name: category.Name,
	})
}

// @Security 		BearerAuth
// @Summary 		List Category
// @Description 	This API for getting categories
// @Tags 			category
// @Produce 		json
// @Accept 			json
// @Param 			page query uint64 true "Page"
// @Param 			limit query uint64 true "Limit"
// @Success			200 {object} models.ListCategory
// @Failure 		404 {object} models.Error
// @Failure 		401 {object} models.Error
// @Failure 		403 {object} models.Error
// @Faulure 		500 {object} models.Error
// @Router 			/v1/categories [GET]
func (h *HandlerV1) ListCategory(c *gin.Context) {
	var (
		jspbMarshal protojson.MarshalOptions
	)
	jspbMarshal.UseProtoNames = true

	ctx, cancel := context.WithTimeout(context.Background(), h.Config.Context.Timeout)
	defer cancel()

	page := c.Query("page")
	limit := c.Query("limit")
	pageInt, err := strconv.Atoi(page)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}
	limitInt, err := strconv.Atoi(limit)
	if err != nil {
		c.JSON(http.StatusBadRequest, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}
	offset := (pageInt - 1) * limitInt

	listCategories, err := h.Service.Category().ListCategory(ctx, &entity.ListReq{
		Offset: offset,
		Limit:  limitInt,
	})
	if err != nil {
		c.JSON(http.StatusNotFound, models.Error{
			Message: err.Error(),
		})
		log.Println(err.Error())
		return
	}

	var categories []models.Category
	for _, category := range listCategories.Category {
		categories = append(categories, models.Category{
			ID:   category.Id,
			Name: category.Name,
		})
	}

	c.JSON(http.StatusOK, models.ListCategory{
		Categories: categories,
		Total:      uint64(listCategories.Totalcount),
	})
}